Refactor: relocate repository interfaces to the repository folder
- Moved all repository interfaces to the repository folder for improved project structure and consistency.
- This change enhances maintainability by grouping repository-related components logically.
