+++ app/src/androidTest/java/com/epfl/beatlink/ui/profile/settings/ChangePasswordScreenTest.kt
@@ -9,7 +9,7 @@ import androidx.compose.ui.test.performClick
 import androidx.compose.ui.test.performScrollTo
 import androidx.compose.ui.test.performTextInput
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.epfl.beatlink.model.auth.FirebaseAuthRepository
+import com.epfl.beatlink.repository.authentication.FirebaseAuthRepository
 import com.epfl.beatlink.ui.navigation.NavigationActions
 import com.epfl.beatlink.viewmodel.auth.FirebaseAuthViewModel
 import io.mockk.mockk
