+++ app/src/androidTest/java/com/epfl/beatlink/ui/profile/settings/SignOutButtonTest.kt
@@ -11,9 +11,9 @@ import androidx.compose.ui.test.performClick
 import androidx.compose.ui.test.performScrollTo
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.epfl.beatlink.model.auth.FirebaseAuthRepository
-import com.epfl.beatlink.model.map.user.MapUserRepository
-import com.epfl.beatlink.model.profile.ProfileRepository
+import com.epfl.beatlink.repository.authentication.FirebaseAuthRepository
+import com.epfl.beatlink.repository.map.user.MapUserRepository
+import com.epfl.beatlink.repository.profile.ProfileRepository
 import com.epfl.beatlink.repository.spotify.auth.SpotifyAuthRepository
 import com.epfl.beatlink.ui.navigation.NavigationActions
 import com.epfl.beatlink.ui.navigation.Screen
