+++ app/src/main/java/com/epfl/beatlink/repository/library/PlaylistRepositoryFirestore.kt
@@ -4,7 +4,6 @@ import android.content.ContentValues.TAG
 import android.graphics.Bitmap
 import android.util.Log
 import com.epfl.beatlink.model.library.Playlist
-import com.epfl.beatlink.model.library.PlaylistRepository
 import com.epfl.beatlink.model.library.PlaylistTrack
 import com.epfl.beatlink.model.spotify.objects.SpotifyTrack
 import com.epfl.beatlink.model.spotify.objects.State
