+++ app/src/main/java/com/epfl/beatlink/repository/spotify/auth/MusicServiceAuthRepository.kt
@@ -1,0 +1,3 @@
+package com.epfl.beatlink.repository.spotify.auth
+
+interface MusicServiceAuthRepository
